Fixed a bug in LinearSolver. Basically, when solving integer inequalities a solution with x >... should have the RHS rounded up. This also called for changing the behavior for solving integer equality comparisons, which haven't been used yet so it doesn't affect anyone. With integer equalities I no longer divide by the coefficient of the LHS, which means that the solver could return an expr like: c*x = ...
Former-commit-id: 9413d9cff42f7901cdb64cc0009eb6fddd6e5226

+++ test/correctness/linear_solve.cpp
@@ -83,9 +83,9 @@ bool check_terms(const std::vector<Term> &expected_terms, const std::vector<Term
 bool test_collect_linear_terms() {
     Var x("x"), y("y"), z("z");
 
-    Expr x_var = Variable::make(Int(32), "x");
-    Expr y_var = Variable::make(Int(32), "y");
-    Expr z_var = Variable::make(Int(32), "z");
+    Expr x_var = Variable::make(Float(32), "x");
+    Expr y_var = Variable::make(Float(32), "y");
+    Expr z_var = Variable::make(Float(32), "z");
 
     /* Simplify an expression by collecting linear terms:
      *
